+++ CHANGELOG.md
@@ -25,3 +25,5 @@ All notable changes to this project will be documented in this file, in reverse 
   possible infinity fread in certain PHP versions.
 - [#28](https://github.com/zendframework/ZendService_Apple_Apns/pull/28) Fixed docblocks
   that prevented proper code completion in some editors.
+- [#29](https://github.com/zendframework/ZendService_Apple_Apns/pull/29) Force
+  TLS vs. SSL due to [Apple moving to TLS](https://developer.apple.com/news/?id=10222014a).